+++ Cargo.toml
@@ -26,3 +26,4 @@ flatgeobuf = { version = "=4.1.0", default-features = false }
 geo-types = { version = "0.7.11", features = ["serde"] }
 geozero = { version = "0.12.0" }
 serde = { version = "1.0.188", features = ["derive"] }
+serde_json = "1.0.135"
