+++ examples/1-constrailt.rs
@@ -1,0 +1,11 @@
+// データ形式寄りの、データ構造
+pub struct MyFeature2_ {
+    pub geometry: geo_types::Point,
+    pub properties: MyFeature2Properties,
+}
+pub struct MyFeature2Properties {
+    pub title: String,
+    pub count: i32,
+}
+
+fn main() {}
