+++ examples/1-generalize-enum.rs
