+++ examples/1-generalize.rs
